Move internal writers to an internal package
